Move MustCompile to a global var
If `regexp.MustCompile()` is inside a handler, it will run every single time that the handler is executed. `regexp.MustCompile()` functions should be ran once when the server starts up for performance reasons. Coworker learned this the hard way when I was helping him figure out why his loop was so slow.